refactor: fold desired-table preparation into the engine, move DesiredTableSource to ports
The desired_tables module held two concerns: the DesiredTableSource
Protocol (the inbound contract user-facing declarations satisfy) and
prepare_desired_tables (a private preparation step with one caller).
The Protocol now lives in ports.py beside the other application
boundary contracts; the preparation function becomes a private helper
in engine.py. Its behaviour tests fold into test_engine.py at the
sync level, where duplicates and empty input were already covered.
